+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### Version 0.0.17
+
+- Auth for Builder is now in progress dialog
+
+### Version 0.0.16
+
+- Fix auth for Builder
+- Change order of tasks
+
 ### Version 0.0.12
 
 - Builder import from Figma improvements